add returning buttons, calculators coins, fix returning value from from call to response, override getData(), add coroutine and exceptionHandler

+++ app/src/main/java/by/khomichenko/testcrypto/domain/models/pair/DataX.kt
@@ -1,7 +1,7 @@
 package by.khomichenko.testcrypto.domain.models.pair
 
 data class DataX(
-    val firstPrice: Int,
+    val firstPrice: Double,
     val high: Double,
     val lastPrice: Double,
     val low: Double,
